feat:Created a new component to show the screes volunteers and educators
